requirements.txt
PyQt5>=5.15.0
google-generativeai>=0.3.0